issue-6958: fastshard - extracting the data structures that can be used across different shard implementations from impl/naive_mirrored to impl/model
