+++ cloud/filestore/libs/storage/fastshard/impl/naive_mirrored/ut/ya.make
@@ -1,8 +1,6 @@
 GTEST()
 
 SRCS(
-    ../persistent_bitmap_ut.cpp
-    ../persistent_hash_table_ut.cpp
     ../shard_ut.cpp
     ../shard_ut_error.cpp
     ../shard_ut_layout.cpp
