+++ cloud/filestore/libs/storage/fastshard/impl/ya.make
@@ -1,4 +1,5 @@
 RECURSE(
     mem
+    model
     naive_mirrored
 )
